requirements.txt
accelerate==0.34.2
altair==5.4.1
beautifulsoup4==4.12.3
blinker==1.8.2
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
filelock==3.16.1
frozendict==2.4.5
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
html5lib==1.1
huggingface-hub==0.25.1
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multitasking==0.0.11
narwhals==1.9.1
networkx==3.3
numpy==2.1.2
packaging==24.1
pandas==2.2.3
peewee==3.17.6
pillow==10.4.0
pipeline==0.1.0
platformdirs==4.3.6
protobuf==5.28.2
psutil==6.0.0
pyarrow==17.0.0
pydeck==0.9.1S
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
safetensors==0.4.5
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
soupsieve==2.6
streamlit==1.39.0
sympy==1.13.3
tenacity==9.0.0
tokenizers==0.20.0
toml==0.10.2
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
transformers==4.45.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==5.0.3
webencodings==0.5.1
yfinance==0.2.44
